Refactor GlobalExceptionHandler
