--- src/main/java/hng_java_boilerplate/squeeze/exceptions/GlobalExceptionHandler.java
@@ -1,38 +1,0 @@
-package hng_java_boilerplate.squeeze.exceptions;
-
-import hng_java_boilerplate.squeeze.util.ResponseMessage;
-import hng_java_boilerplate.squeeze.util.ValidationErrorResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<ValidationErrorResponse> handleValidationExceptions(MethodArgumentNotValidException ex) {
-        List<ValidationErrorResponse.FieldError> errors = ex.getBindingResult()
-                .getFieldErrors()
-                .stream()
-                .map(err -> new ValidationErrorResponse.FieldError(err.getField(), err.getDefaultMessage()))
-                .collect(Collectors.toList());
-
-        ValidationErrorResponse response = new ValidationErrorResponse("There are missing field(s) in the form", errors);
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(DuplicateEmailException.class)
-    public ResponseEntity<ResponseMessage> handleDuplicateEmailException(DuplicateEmailException ex) {
-        return ResponseEntity.status(HttpStatus.CONFLICT).body(new ResponseMessage(ex.getMessage(), HttpStatus.CONFLICT.value()));
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ResponseMessage> handleGeneralException(Exception ex) {
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ResponseMessage("An error occurred", HttpStatus.INTERNAL_SERVER_ERROR.value()));
-    }
-}
